Todos los errores 404

+++ src/routes/albums/delAlbum.js
@@ -2,10 +2,21 @@ import { prisma } from '../../db/index.js';
 
 export const delAlbum = async (ctx) => {
   // elimino album
-  const deleteAlbum = await prisma.album.delete({
+  console.log(ctx.params);
+  const album = await prisma.album.findUnique({
     where: {
       id: ctx.params.albumId,
     },
   });
-  ctx.status = 204;
+  if (album) {
+    const deleteAlbum = await prisma.album.delete({
+      where: {
+        id: ctx.params.albumId,
+      },
+    });
+    ctx.status = 204;
+  } else {
+    ctx.body = 'No existe el album';
+    ctx.status = 404;
+  }
 };
