+++ src/routes/artists/delArtist.js
@@ -2,10 +2,20 @@ import { prisma } from '../../db/index.js';
 
 export const delArtist = async (ctx) => {
   // elimino album
-  const deleteArtist = await prisma.artist.delete({
+  const artist = await prisma.artist.findUnique({
     where: {
       id: ctx.params.artistId,
     },
   });
-  ctx.status = 204;
+  if (artist) {
+    const deleteArtist = await prisma.artist.delete({
+      where: {
+        id: ctx.params.artistId,
+      },
+    });
+    ctx.status = 204;
+  } else {
+    ctx.body = 'No existe el artista';
+    ctx.status = 404;
+  }
 };
