+++ src/routes/tracks/delTrack.js
@@ -1,10 +1,20 @@
 import { prisma } from '../../db/index.js';
 
 export const delTrack = async (ctx) => {
-  const deleteTrack = await prisma.track.delete({
+  const track = await prisma.track.findUnique({
     where: {
       id: ctx.params.trackId,
     },
   });
-  ctx.status = 204;
+  if (track) {
+    const deleteTrack = await prisma.track.delete({
+      where: {
+        id: ctx.params.trackId,
+      },
+    });
+    ctx.status = 204;
+  } else {
+    ctx.body = 'No existe el track';
+    ctx.status = 404;
+  }
 };
